moved credits code to a credits module
