+++ src/main.rs
@@ -1,10 +1,6 @@
 extern crate sdl2;
 
-use std::time::Duration;
-use std::thread;
-//use sdl2::rect::Rect;
-use sdl2::pixels::Color;
-use sdl2::image::LoadTexture;
+mod credits;
 
 fn main() {
 	let sdl_cxt = sdl2::init().unwrap();
@@ -17,71 +13,9 @@ fn main() {
 
 	let wincan = window.into_canvas().accelerated();
 
-	let mut wincan = wincan.build()
+	let wincan = wincan.build()
 		.map_err(|e| e.to_string())
 		.unwrap();
 
-//	let event_pump = sdl_cxt.event_pump().unwrap();
-
-
-
-	let texture_creator = wincan.texture_creator();
-
-	let andrew = texture_creator.load_texture("assets/credits/andrew_credits.png").unwrap();
-	let kira = texture_creator.load_texture("assets/credits/kira_credits.jpg").unwrap();
-	let josh = texture_creator.load_texture("assets/credits/josh_credits.png").unwrap();
-	let alvyn = texture_creator.load_texture("assets/credits/alvyn_credits.png").unwrap();
-	let greg = texture_creator.load_texture("assets/credits/greg_credits.jpg").unwrap();
-	let jake = texture_creator.load_texture("assets/credits/jake_credits.jpeg").unwrap();
-	let bryce = texture_creator.load_texture("assets/credits/bryce_credits.png").unwrap();
-	let austin = texture_creator.load_texture("assets/credits/austin_credits.png").unwrap();
-	let evan = texture_creator.load_texture("assets/credits/evan_credits.png").unwrap();
-
-	wincan.set_draw_color(Color::RGBA(0, 128, 128, 255));
-	wincan.clear();
-
-	wincan.copy(&andrew, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500));
-
-	wincan.copy(&kira, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500));
-
-	wincan.copy(&josh, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500));
-
-	wincan.copy(&alvyn, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500));
-
-	wincan.copy(&greg, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500));
-
-	wincan.copy(&jake, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500));
-
-	wincan.copy(&bryce, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500)); 
-
-	wincan.copy(&austin, None, None).ok();
-	wincan.present();
-
-	thread::sleep(Duration::from_millis(1500));
-
-	wincan.copy(&evan, None, None).ok();
-	wincan.present();
-	
-	thread::sleep(Duration::from_millis(1500));
+	credits::show_credits(wincan);
 }
